@@ -1,14 +1,12 @@
-'use strict'
+import bench from 'fastbench'
+import fs from 'fs'
+import { createLogger, JsonFormat } from '../lib/Apheleia.mjs'
+import pino from 'pino'
+import SonicBoom from 'sonic-boom'
+import winston from 'winston'
 
-const bench = require('fastbench')
-const fs = require('fs')
-const { createLogger, JsonFormat } = require('../lib/Apheleia')
-const pino = require('pino')
-const SonicBoom = require('sonic-boom')
-const winston = require('winston')
-
-const pinoLogger = pino(fs.createWriteStream('/dev/null')).child({ module: 'test' })
-const pinoExtreme = pino(pino.destination({ sync: false, dest: '/dev/null' })).child({ module: 'test' })
+const pinoLogger = pino(fs.createWriteStream('/dev/null'))
+const pinoExtreme = pino(pino.destination({ sync: false, dest: '/dev/null' }))
 
 const winstonLogger = winston.createLogger({
   transports: [
@@ -17,53 +15,62 @@ const winstonLogger = winston.createLogger({
       format: winston.format.json()
     })
   ]
-}).child({ module: 'test' })
+})
 
 const apheleiaLoggerJson = createLogger({
   stream: fs.createWriteStream('/dev/null'),
   format: new JsonFormat()
-}).child({ module: 'test' })
+})
 
 const apheleiaLoggerJsonSonicBoom = createLogger({
   stream: new SonicBoom({ dest: '/dev/null', minLength: 0, sync: false }),
   format: new JsonFormat()
-}).child({ module: 'test' })
+})
 
-module.exports = function childLogging () {
+const testObject = {
+  ip: '127.0.0.1',
+  id: 'CFPouPXNLMFLtrfVxfzW',
+  method: 'GET',
+  protocol: 'http:',
+  url: 'http://127.0.0.1:3000/test',
+  query: '?test=true'
+}
+
+export default function objectJsonFormat () {
   return new Promise((resolve) => {
     const run = bench([
       function benchWinston (cb) {
         for (let i = 0; i < 10; i++) {
-          winstonLogger.info('hello world')
+          winstonLogger.info('hello world', testObject)
         }
         setImmediate(cb)
       },
       function benchPino (cb) {
         for (let i = 0; i < 10; i++) {
-          pinoLogger.info('hello world')
+          pinoLogger.info('hello world', testObject)
         }
         setImmediate(cb)
       },
       function benchApheleiaJson (cb) {
         for (let i = 0; i < 10; i++) {
-          apheleiaLoggerJson.info('hello world')
+          apheleiaLoggerJson.info('hello world', testObject)
         }
         setImmediate(cb)
       },
       function benchPinoExtreme (cb) {
         for (let i = 0; i < 10; i++) {
-          pinoExtreme.info('hello world')
+          pinoExtreme.info('hello world', testObject)
         }
         setImmediate(cb)
       },
       function benchApheleiaJsonSonicBoom (cb) {
         for (let i = 0; i < 10; i++) {
-          apheleiaLoggerJsonSonicBoom.info('hello world')
+          apheleiaLoggerJsonSonicBoom.info('hello world', testObject)
         }
         setImmediate(cb)
       }
     ], 100000)
-    console.log('\nChild Logging:\n')
+    console.log('\nObject JsonFormat:\n')
     run(resolve)
   })
 }
