@@ -1,25 +1,24 @@
-'use strict'
-
-const DateFormat = require('fast-date-format')
-const {
+import { DateFormatter } from 'fast-date-format'
+import {
   DEFAULT_LEVELS,
   propertiesToIgnore
-} = require('./Defaults')
-const jsonEscape = require('json-escaping')
-const { dateFormatterSym } = require('./Symbols')
-const fastifySerializers = require('./FastifySerializers').json
+} from './Defaults.mjs'
+import jsonEscape from 'json-escaping'
+import { serialzeRequestJson, serialzeResponseJson } from './FastifySerializers.mjs'
 
-module.exports = class JsonFormat {
+export class JsonFormat {
+  #dateFormatter
+
   constructor (options = {}) {
-    this[dateFormatterSym] = new DateFormat({
+    this.#dateFormatter = new DateFormatter({
       dateFormat: options.dateFormat || 'x',
       cache: options.fastTimestamp
     })
-    this.serializers = options.serializers || fastifySerializers
+    this.serializers = options.serializers || { req: serialzeRequestJson, res: serialzeResponseJson }
   }
 
   transform (level, message, args, meta = '') {
-    const timestamp = this[dateFormatterSym].format()
+    const timestamp = this.#dateFormatter.format()
     let line = `{"time":${timestamp},"level":${DEFAULT_LEVELS[level]}`
 
     let stack = null
