@@ -1,13 +1,11 @@
-'use strict'
+import { Logger } from './Logger.mjs'
+import { SimpleFormat } from './SimpleFormat.mjs'
+import { JsonFormat } from './JsonFormat.mjs'
+import { Transport } from './Transport.mjs'
+import { ConsoleTransport } from './ConsoleTransport.mjs'
+import { DEFAULT_LEVEL } from './Defaults.mjs'
 
-const Logger = require('./Logger')
-const SimpleFormat = require('./SimpleFormat')
-const JsonFormat = require('./JsonFormat')
-const Transport = require('./Transport')
-const ConsoleTransport = require('./ConsoleTransport')
-const { DEFAULT_LEVEL } = require('./Defaults')
-
-function createLogger (options = {}) {
+export function createLogger (options = {}) {
   options.levelString = options.level || DEFAULT_LEVEL
   options.transports = options.transports || [getTransport(options)]
   return new Logger(options)
@@ -23,8 +21,7 @@ function getTransport (options) {
   }
 }
 
-module.exports = {
-  createLogger,
+export {
   Transport,
   SimpleFormat,
   JsonFormat,
