@@ -1,17 +1,13 @@
-'use strict'
+import { Transport } from './Transport.mjs'
+import { EOL } from 'os'
+import { formatSym } from './Symbols.mjs'
 
-const Transport = require('./Transport')
+export class ConsoleTransport extends Transport {
+  #stderrLevels
 
-const {
-  formatSym,
-  stderrLevelsSym
-} = require('./Symbols')
-const { EOL } = require('os')
-
-module.exports = class ConsoleTransport extends Transport {
   constructor (options) {
     super({ format: options.format })
-    this[stderrLevelsSym] = options.stderrLevels || {
+    this.#stderrLevels = options.stderrLevels || {
       WARN: true,
       ERROR: true,
       FATAL: true
@@ -20,7 +16,7 @@ module.exports = class ConsoleTransport extends Transport {
 
   log (level, message, args, meta) {
     const line = this[formatSym].transform(level, message, args, meta)
-    if (this[stderrLevelsSym][level]) {
+    if (this.#stderrLevels[level]) {
       process.stderr.write(`${line}${EOL}`)
     } else {
       process.stdout.write(`${line}${EOL}`)
