@@ -1,5 +1,3 @@
-'use strict'
-
 const DEFAULT_LEVELS = {
   TRACE: 10,
   DEBUG: 20,
@@ -15,7 +13,7 @@ const propertiesToIgnore = [
   'genReqId'
 ]
 
-module.exports = {
+export {
   DEFAULT_LEVELS,
   DEFAULT_LEVEL,
   propertiesToIgnore
