@@ -1,17 +1,14 @@
-'use strict'
+import { DateFormatter } from 'fast-date-format'
+import { serialzeRequestSimple, serialzeResponseSimple } from './FastifySerializers.mjs'
+import { propertiesToIgnore } from './Defaults.mjs'
 
-const DateFormat = require('fast-date-format')
-const fastifySerializers = require('./FastifySerializers').simple
-const { propertiesToIgnore } = require('./Defaults')
-const { formatSym, formatObjectSym } = require('./Symbols')
-
-module.exports = class SimpleFormat {
+export class SimpleFormat {
   constructor (options = {}) {
-    this.dateFormatter = new DateFormat({
+    this.dateFormatter = new DateFormatter({
       dateFormat: options.dateFormat || 'YYYY-MM-DDTHH:mm:ss.SSS[Z]',
       cache: options.fastTimestamp
     })
-    this.serializers = options.serializers || fastifySerializers
+    this.serializers = options.serializers || { req: serialzeRequestSimple, res: serialzeResponseSimple }
   }
 
   transform (level, message, args, meta = '') {
@@ -26,7 +23,7 @@ module.exports = class SimpleFormat {
         stack = '\n' + arg.stack
         continue
       }
-      line += this[formatSym](arg)
+      line += this.#format(arg)
     }
     return `${date} ${level} ${message}${line}${meta}${stack}`
   }
@@ -53,10 +50,10 @@ module.exports = class SimpleFormat {
     return ` ${key}=${value}`
   }
 
-  [formatSym] (arg) {
+  #format (arg) {
     switch (typeof arg) {
       case 'object':
-        return this[formatObjectSym](arg)
+        return this.#formatObject(arg)
       case 'function':
         return ''
       default:
@@ -64,7 +61,7 @@ module.exports = class SimpleFormat {
     }
   }
 
-  [formatObjectSym] (arg) {
+  #formatObject (arg) {
     let formatted = ''
     for (const key in arg) {
       if (key === 'level' || key === 'message') {
