@@ -1,20 +1,17 @@
-'use strict'
+import { EOL } from 'os'
+import { formatSym } from './Symbols.mjs'
 
-const {
-  formatSym,
-  streamSym
-} = require('./Symbols')
-const { EOL } = require('os')
+export class Transport {
+  #stream
 
-module.exports = class Transport {
   constructor (options) {
-    this[streamSym] = options.stream
+    this.#stream = options.stream
     this[formatSym] = options.format
   }
 
   log (level, message, args, meta) {
     const line = this[formatSym].transform(level, message, args, meta)
-    this[streamSym].write(`${line}${EOL}`)
+    this.#stream.write(`${line}${EOL}`)
   }
 
   formatMetaDataObject (meta) {
@@ -27,8 +24,8 @@ module.exports = class Transport {
 
   end () {
     return new Promise((resolve) => {
-      this[streamSym].on('finish', resolve)
-      this[streamSym].end()
+      this.#stream.on('finish', resolve)
+      this.#stream.end()
     })
   }
 }
